Sickness class creation and Order class medification
One of those ones... Patients are united, they have acute and periodic sicknesses. Sicknesses contain which medication is needed for them and at what frequency. For this feature medication now has many new variables that determine their consumption frequency and how many the consumer has got in hand of them. This will need a lot of revisiting, though it is a good architecture.